# swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/solutions": {
      "get": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Get available solutions list.",
        "operationId": "GetSolutionsList",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "available solutions",
            "schema": {
              "$ref": "#/definitions/AvailableSolutionsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "AvailableSolutions"
        ],
        "summary": "Add available solution.",
        "operationId": "AddTemplate",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AvailableSolution"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "solution added"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/user_solutions": {
      "get": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Get running solutions list.",
        "operationId": "GetSolutionsList",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "running solutions list",
            "schema": {
              "$ref": "#/definitions/UserSolutionsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Run solution.",
        "operationId": "RunSolution",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserSolution"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "solution created",
            "schema": {
              "$ref": "#/definitions/RunSolutionResponce"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/user_solutions/{solution}": {
      "delete": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Delete solution.",
        "operationId": "DeleteSolution",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "solution deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/user_solutions/{solution}/deployments": {
      "get": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Get solution deployments.",
        "operationId": "GetSolutionsDeployments",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "solution deployments",
            "schema": {
              "$ref": "#/definitions/DeploymentsList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/user_solutions/{solution}/services": {
      "get": {
        "tags": [
          "UserSolutions"
        ],
        "summary": "Get solution services.",
        "operationId": "GetSolutionsServices",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "solutions services",
            "schema": {
              "$ref": "#/definitions/ServicesList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    }
  },
  "definitions": {
    "AvailableSolution": {
      "description": "AvailableSolution -- solution which user can run",
      "type": "object",
      "properties": {
        "Active": {
          "type": "boolean"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Images"
        },
        "limits": {
          "$ref": "#/definitions/Limits"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "url": {
          "type": "string",
          "x-go-name": "URL"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
    "AvailableSolutionsList": {
      "description": "AvailableSolutionsList -- list of available solutions",
      "type": "object",
      "properties": {
        "solutions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AvailableSolution"
          },
          "x-go-name": "Solutions"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "Container": {
      "description": "Container -- model for container in deployment",
      "type": "object",
      "required": [
        "image",
        "name",
        "limits"
      ],
      "properties": {
        "commands": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Commands"
        },
        "config_maps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerVolume"
          },
          "x-go-name": "ConfigMaps"
        },
        "env": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Env"
          },
          "x-go-name": "Env"
        },
        "image": {
          "type": "string",
          "x-go-name": "Image"
        },
        "limits": {
          "$ref": "#/definitions/Resource"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerPort"
          },
          "x-go-name": "Ports"
        },
        "volume_mounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ContainerVolume"
          },
          "x-go-name": "VolumeMounts"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ContainerPort": {
      "description": "ContainerPort -- model for port in container",
      "type": "object",
      "required": [
        "name",
        "port",
        "protocol"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Port"
        },
        "protocol": {
          "$ref": "#/definitions/Protocol"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ContainerVolume": {
      "description": "ContainerVolume -- volume (or config map) mounted in container",
      "type": "object",
      "required": [
        "name",
        "mount_path"
      ],
      "properties": {
        "mode": {
          "type": "string",
          "x-go-name": "Mode"
        },
        "mount_path": {
          "type": "string",
          "x-go-name": "MountPath"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "sub_path": {
          "type": "string",
          "x-go-name": "SubPath"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "DeploymentStatus": {
      "description": "DeploymentStatus -- kubernetes status of deployment",
      "type": "object",
      "properties": {
        "available_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AvailableReplicas"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "ready_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ReadyReplicas"
        },
        "replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Replicas"
        },
        "unavailable_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UnavailableReplicas"
        },
        "updated_at": {
          "description": "update date in RFC3339 format",
          "type": "string",
          "x-go-name": "UpdatedAt"
        },
        "updated_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdatedReplicas"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "DeploymentWithOwner": {
      "description": "DeploymentWithOwner -- model for deployment with owner",
      "type": "object",
      "required": [
        "containers",
        "name",
        "replicas",
        "owner"
      ],
      "properties": {
        "containers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Container"
          },
          "x-go-name": "Containers"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Replicas"
        },
        "status": {
          "$ref": "#/definitions/DeploymentStatus"
        },
        "total_cpu": {
          "description": "total CPU usage by all containers in this deployment",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TotalCPU"
        },
        "total_memory": {
          "description": "total RAM usage by all containers in this deployment",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TotalMemory"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/pkg/model"
    },
    "DeploymentsList": {
      "description": "DeploymentsList -- model for deployments list",
      "type": "object",
      "properties": {
        "deployments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DeploymentWithOwner"
          },
          "x-go-name": "Deployments"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/pkg/model"
    },
...
    "Env": {
      "description": "Env -- key-value pair of enviroment variables",
      "type": "object",
      "required": [
        "value",
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "value": {
          "type": "string",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
...
    "Limits": {
      "description": "Limits -- solution resources limits",
      "type": "object",
      "properties": {
        "cpu": {
          "type": "string",
          "x-go-name": "CPU"
        },
        "ram": {
          "type": "string",
          "x-go-name": "RAM"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "Protocol": {
      "description": "represents port protocol type (TCP or UDP)",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "Resource": {
      "description": "Resource -- represents namespace CPU and RAM",
      "type": "object",
      "required": [
        "cpu",
        "memory"
      ],
      "properties": {
        "cpu": {
          "description": "CPU in m",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CPU"
        },
        "memory": {
          "description": "RAM in Mi",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Memory"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "RunSolutionResponce": {
      "description": "RunSolutionResponce -- responce to run solution request",
      "type": "object",
      "properties": {
        "created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Errors"
        },
        "not_created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "NotCreated"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
    "ServicePort": {
      "description": "represent service port",
      "type": "object",
      "required": [
        "name",
        "target_port",
        "protocol"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Port"
        },
        "protocol": {
          "$ref": "#/definitions/Protocol"
        },
        "target_port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TargetPort"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ServiceWithOwner": {
      "description": "ServiceWithOwner -- model for service with owner",
      "type": "object",
      "required": [
        "name",
        "ports",
        "owner"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deploy": {
          "type": "string",
          "x-go-name": "Deploy"
        },
        "domain": {
          "type": "string",
          "x-go-name": "Domain"
        },
        "hidden": {
          "description": "hide service from users",
          "type": "boolean",
          "x-go-name": "Hidden"
        },
        "ips": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "IPs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "no_selector": {
          "description": "don't add selectors to service (so don't create endpoint)",
          "type": "boolean",
          "x-go-name": "NoSelector"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServicePort"
          },
          "x-go-name": "Ports"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/pkg/model"
    },
    "ServicesList": {
      "description": "ServicesList -- model for services list",
      "type": "object",
      "properties": {
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceWithOwner"
          },
          "x-go-name": "Services"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/pkg/model"
    },
...
    "UserSolution": {
      "description": "UserSolution -- running solution",
      "type": "object",
      "required": [
        "template",
        "name",
        "namespace"
      ],
      "properties": {
        "branch": {
          "type": "string",
          "x-go-name": "Branch"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Env"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "template": {
          "type": "string",
          "x-go-name": "Template"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
    "UserSolutionsList": {
      "description": "UserSolutionsList -- list of running solution",
      "type": "object",
      "properties": {
        "solutions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserSolution"
          },
          "x-go-name": "Solutions"
        }
      },
      "x-go-package": "git.containerum.net/ch/solutions/pkg/models"
    },
...
  },
  "parameters": {
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}